feat: day11
Drew a lot on what I learned in day9 on managing grids. I'm still
not in love with the way I'm checking grid coordinates (all those
bounds checks to make sure I'm not wrapping off an edge is a bummer)
but wrapping successive iterations in an iterator is a neat idea.

I wonder if I'll look back at this in a year or so and think:

"Man, remember when I thought iterators mutating state was a good
idea?"