feat: day13
A refreshing puzzle. This one was pretty approachable on its face
and while it took me a bit to figure out the formula for "Hey,
where should the folded point go" -- and indeed I came up with the
wrong one at first -- it became clear once I had to impl Display
for Grid.