feat: day14
This one was a toughie. The naive implementation got me through part1
but trying to calculate part2 even for the test case took over 11hours.

The new counter-based implementation took ~4s.

I had to muck with the logic for PolymerPairCounter.perform_insertion
a lot to make it do what I wanted it to do. Notably, Entry.or_insert
seems to evaluate no matter what. Since there was a reasonable panic
in the .or_insert path (trying to subtract when there was no value),
it threw errors for a long time before I realized that if I used
Entry.or_insert_with the closure is only called if it needs to be.